# apps/api/docs/swagger.yaml
definitions:
...
  portfolio.CreateManualPortfolioTransactionRequest:
    properties:
      amount:
        type: string
      description:
        type: string
      listing_id:
        type: string
      occurred_at:
        type: string
      quantity:
        type: string
      type:
        type: string
      vendor_id:
        type: string
    type: object
  portfolio.ManualPortfolioTransactionResponse:
    properties:
      account_id:
        type: string
      amount:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      isin:
        type: string
      listing_id:
        type: string
      occurred_at:
        type: string
      origin:
        type: string
      quantity:
        type: string
      source:
        type: string
      symbol:
        type: string
      type:
        type: string
      unit_price:
        type: string
      updated_at:
        type: string
    type: object
  portfolio.PaginationResponse:
    properties:
      count:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  portfolio.PortfolioTransactionResponse:
    properties:
      account_id:
        type: string
      amount:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      isin:
        type: string
      listing_id:
        type: string
      occurred_at:
        type: string
      origin:
        type: string
      quantity:
        type: string
      source:
        type: string
      symbol:
        type: string
      type:
        type: string
      unit_price:
        type: string
      updated_at:
        type: string
    type: object
  portfolio.PortfolioTransactionsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/portfolio.PortfolioTransactionResponse'
        type: array
      pagination:
        $ref: '#/definitions/portfolio.PaginationResponse'
    type: object
...
info:
  contact: {}
paths:
...
  /portfolio/transactions:
    get:
      consumes:
      - application/json
      description: Returns portfolio transactions for the given account and optional
        date range.
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      - description: Page size (10, 25, 50, 100)
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: 'Sort field: date'
        in: query
        name: sort_by
        type: string
      - description: 'Sort order: asc or desc'
        in: query
        name: sort_order
        type: string
      - description: Contains search over description, source, symbol, isin
        in: query
        name: q
        type: string
      - description: 'Transaction type: BUY, SELL, DIVIDEND, TAX, FEE, CASH'
        in: query
        name: type
        type: string
      - description: 'Transaction origin: IMPORT, MANUAL'
        in: query
        name: origin
        type: string
      - description: Case-insensitive contains filter on source
        in: query
        name: source
        type: string
      - description: Case-insensitive contains filter on symbol/isin
        in: query
        name: listing
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/portfolio.PortfolioTransactionsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get portfolio transactions
      tags:
      - portfolio
  /portfolio/transactions/manual:
    post:
      consumes:
      - application/json
      description: Creates a manual portfolio transaction and persists it without
        publishing rebuild events.
      parameters:
      - description: Manual portfolio transaction payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/portfolio.CreateManualPortfolioTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/portfolio.ManualPortfolioTransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create manual portfolio transaction
      tags:
      - portfolio
...
swagger: "2.0"